requirements.txt
pynput==1.7.6
pyserial==3.5
